# docs/openapi.yaml
openapi: 3.0.3
info:
  title: GeoAgentic Emergency Response System Backend API
  description: >
    Intelligent decision-support and dispatch platform for monitoring emergency vehicles,
    detecting route deviations, predicting delays, running Gemini AI advisory analysis,
    and evaluating authoritative operational decisions.
  version: 1.0.0
servers:
  - url: http://localhost:5001/api
    description: Local Development Server
paths:
...
  /auth/register:
    post:
      summary: Register a new operator/user account
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name:
                  type: string
                  example: Jane Doe
                email:
                  type: string
                  format: email
                  example: operator@geoagent.local
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: SecurePass123!
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  user: { $ref: '#/components/schemas/SafeUser' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /auth/login:
    post:
      summary: Authenticate user and issue JWT cookie
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email, example: operator@geoagent.local }
                password: { type: string, format: password, example: SecurePass123! }
      responses:
        '200':
          description: Successfully authenticated
          headers:
            Set-Cookie:
              schema: { type: string, example: token=jwt_token_here; Path=/; HttpOnly; SameSite=Strict }
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  token: { type: string }
                  user: { $ref: '#/components/schemas/SafeUser' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /routes:
    post:
      summary: Generate and persist a planned route
      tags: [Routes]
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [emergencyId, vehicleId, origin, destination]
              properties:
                emergencyId: { type: string }
                vehicleId: { type: string }
                origin: { $ref: '#/components/schemas/GeoJSONPoint' }
                destination: { $ref: '#/components/schemas/GeoJSONPoint' }
                routeType: { type: string, enum: [PLANNED, ALTERNATIVE, HISTORICAL], default: PLANNED }
      responses:
        '201':
          description: Route created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { $ref: '#/components/schemas/Route' }
...
  /admin/routes:
    get:
      summary: List planned and alternative routes (ADMIN only)
      tags: [Admin]
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        '200':
          description: Paginated route paths
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    GeoJSONPoint:
      type: object
      required: [type, coordinates]
      properties:
        type: { type: string, enum: [Point], example: Point }
        coordinates:
          type: array
          items: { type: number }
          minItems: 2
          maxItems: 2
          example: [77.5946, 12.9716]
          description: "[longitude, latitude] in WGS84"
    GeoJSONLineString:
      type: object
      required: [type, coordinates]
      properties:
        type: { type: string, enum: [LineString], example: LineString }
        coordinates:
          type: array
          items:
            type: array
            items: { type: number }
            minItems: 2
            maxItems: 2
          example: [[77.5946, 12.9716], [77.6000, 12.9750]]
    SafeUser:
      type: object
      properties:
        id: { type: string }
        name: { type: string, example: Jane Doe }
        email: { type: string, format: email, example: operator@geoagent.local }
        role: { type: string, enum: [ADMIN, CONTROL_ROOM, DRIVER, PARAMEDIC] }
...
    Route:
      type: object
      properties:
        id: { type: string }
        routeId: { type: string, example: ROUTE-1001 }
        emergency: { type: string }
        vehicle: { type: string }
        origin: { $ref: '#/components/schemas/GeoJSONPoint' }
        destination: { $ref: '#/components/schemas/GeoJSONPoint' }
        geometry: { $ref: '#/components/schemas/GeoJSONLineString' }
        distance: { type: number, example: 5896, description: Distance in meters }
        duration: { type: number, example: 720, description: Duration in seconds }
        status: { type: string, enum: [ACTIVE, COMPLETED, ABANDONED] }
...
  responses:
    BadRequest:
      description: Bad Request (validation failure or tampering)
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: false }
              error: { type: string, example: Validation error message }
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: false }
              error: { type: string, example: Authentication required }
    Forbidden:
      description: Forbidden (insufficient permissions)
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: false }
              error: { type: string, example: Forbidden - Insufficient privileges }
...
    Conflict:
      description: Conflict (duplicate unique key)
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: false }
              error: { type: string, example: Email is already registered }